POC for pluggable event handlers
